+++ CHANGELOG.rst
@@ -12,8 +12,14 @@ The semantic versioning only considers the public API as described in
 paths are considered internals and can change in minor and patch releases.
 
 
-v4.26.3 (2023-11-??)
+v4.27.0 (2023-11-??)
 --------------------
+
+Added
+^^^^^
+- Support for pydantic's ``SecretStr`` type.
+- New ``SecretStr`` type in ``jsonargparse.typing`` to provide secret support
+  without additional dependency.
 
 Fixed
 ^^^^^